requirements.txt
discord.py
sqlalchemy
pendulum
requests